refactor: 앱 색상 변경
