+++ ShoppingPractice/Presentation/Base/BaseButton.swift
@@ -1,0 +1,19 @@
+//
+//  BaseButton.swift
+//  ShoppingPractice
+//
+//  Created by BAE on 2/7/25.
+//
+
+import UIKit
+
+class BaseButton: UIButton {
+    override init(frame: CGRect) {
+        super.init(frame: frame)
+    }
+    
+    @available(*, unavailable)
+    required init?(coder: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
+    }
+}
