+++ ShoppingPractice/Presentation/Base/BaseCollectionViewCell.swift
@@ -10,21 +10,14 @@ import UIKit
 class BaseCollectionViewCell: UICollectionViewCell {
     override init(frame: CGRect) {
         super.init(frame: frame)
+        backgroundColor = .secondarySystemGroupedBackground
+        configCell()
     }
     
     @available(*, unavailable)
     required init?(coder: NSCoder) {
         fatalError("init(coder:) has not been implemented")
     }
-}
-
-class BaseButton: UIButton {
-    override init(frame: CGRect) {
-        super.init(frame: frame)
-    }
     
-    @available(*, unavailable)
-    required init?(coder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
+    func configCell() { }
 }
